+++ muikku-atests/src/test/java/fi/otavanopisto/muikku/wcag/AbstractWCAGTest.java
@@ -1,16 +1,9 @@
 package fi.otavanopisto.muikku.wcag;
 
-import static org.junit.Assert.assertTrue;
-
 import java.net.URL;
-import java.util.HashMap;
-import java.util.Map;
 
-import org.json.JSONArray;
 import org.junit.After;
 import org.junit.Before;
-
-import com.deque.axe.AXE;
 
 import fi.otavanopisto.muikku.ui.AbstractUITest;
 
@@ -24,31 +17,8 @@ public class AbstractWCAGTest extends AbstractUITest {
   }
   
   @After
-  protected void reportWCAG() {
-    if (!System.getProperty("it.profile").equals("sauce-it")) {
-      if (this.violationList != null) {
-        if (!this.violationList.isEmpty()) {
-          String violationsString = "";          
-          for (Map.Entry<String, JSONArray> violation : violationList.entrySet()) {
-            violationsString += System.getProperty("line.separator");
-            violationsString += violation.getKey();
-            violationsString += System.getProperty("line.separator");
-            violationsString += AXE.report(violation.getValue());
-            violationsString += System.getProperty("line.separator");
-          }
-          assertTrue(violationsString, false);
-        }
-      }
-    }
+  protected void report() {
+    reportWCAG();
   }
-    
-  protected void testAccessibility() {
-    if (!System.getProperty("it.profile").equals("sauce-it")) {
-      if (this.violationList == null) {
-        this.violationList = new HashMap<String, JSONArray>();
-      }
-      this.violationList.put("default", new AXE.Builder(getWebDriver(), scriptUrl).analyze().getJSONArray("violations"));
-    }
-  }
-  
+
 }
